backend/flask-service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Cors==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Flask-JWT-Extended==4.3.1
Flask-SocketIO==5.3.6
pytest==7.0.1
PyJWT==2.8.0
bcrypt==4.0.1
gunicorn==21.2.0
python-jose==3.3.0
email-validator==2.0.0.post2
redis==5.0.0
bleach==6.0.0 